refactor API enums
