+++ ui/src/App.tsx
@@ -12,7 +12,7 @@ import { FacilityPage } from './components/FacilityPage/FacilityPage'
 import { isDevOrTestEnv } from './env'
 import { handleCheckForMissingTranslations } from './utils'
 import { ReleaseSearch } from './components/ReleaseSearch/ReleaseSearch'
-import { Medium } from './api/models/Medium'
+import { Medium } from './api/enums/Medium'
 import { ScrollToTop } from './components/ScrollToTop'
 import { Footer } from './components/Footer'
 import { WasteTransferSearch } from './components/WasteTransferSearch/WasteTransferSearch'
