+++ ui/src/api/enums/WasteClassificationCode.tsx
