+++ ui/src/api/enums/WasteTreatmentCode.tsx
