+++ ui/src/components/ReleaseSearch/ReleaseFilterPanel.tsx
@@ -5,11 +5,11 @@ import { FormEvent, useEffect, useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useHistory, useLocation } from 'react-router-dom'
 import styled from 'styled-components'
-import { PollutantCode } from '../../api/models/PollutantCode'
+import { PollutantCode } from '../../api/enums/PollutantCode'
 import { OptionType } from '../../models/OptionType'
 import { URLSearchParamName } from '../../models/URLSearchParamName'
 import { ChakraSelect } from '../ChakraReactSelect'
-import { Medium } from '../../api/models/Medium'
+import { Medium } from '../../api/enums/Medium'
 import { getLongPollutantLabel, getPollutantLabel } from '../../utils'
 import { asOption } from '../../models/OptionType'
 import { TranslationKeys } from '../../react-i18next'
